test(projection-typeorm): hoist createRollForwardEventBasedOn util from storeNftMetadata.test.ts
