+++ packages/projection-typeorm/test/operators/storeNftMetadata.test.ts
@@ -15,13 +15,14 @@ import {
   typeormTransactionCommit,
   withTypeormTransaction
 } from '../../src';
-import { BaseProjectionEvent, Bootstrap, Mappers, ProjectionEvent, requestNext } from '@cardano-sdk/projection';
+import { Bootstrap, Mappers, ProjectionEvent, requestNext } from '@cardano-sdk/projection';
 import { ChainSyncDataSet, chainSyncData, generateRandomHexString, logger } from '@cardano-sdk/util-dev';
 import { Observable, defer, firstValueFrom, from, lastValueFrom, toArray } from 'rxjs';
 import { QueryRunner, Repository } from 'typeorm';
 import {
   createProjectorTilFirst,
   createRollBackwardEventFor,
+  createRollForwardEventBasedOn,
   createStubBlockHeader,
   createStubProjectionSource,
   createStubRollForwardEvent
@@ -123,29 +124,6 @@ const patchNftMetadataNameCip68InTx = (
     }
   }
   return tx;
-};
-
-const createRollForwardEventBasedOn = (
-  evt: BaseProjectionEvent,
-  patchBlock: (block: Cardano.Block) => Cardano.Block
-): BaseProjectionEvent => {
-  const updateBlockHeader: Cardano.PartialBlockHeader = {
-    blockNo: Cardano.BlockNo(evt.block.header.blockNo + 1),
-    hash: Cardano.BlockId(generateRandomHexString(64)),
-    slot: Cardano.Slot(evt.block.header.slot + 20)
-  };
-  return {
-    block: {
-      ...patchBlock(evt.block),
-      header: updateBlockHeader
-    },
-    crossEpochBoundary: false,
-    epochNo: evt.epochNo,
-    eraSummaries: evt.eraSummaries,
-    eventType: ChainSyncEventType.RollForward,
-    genesisParameters: evt.genesisParameters,
-    tip: updateBlockHeader
-  };
 };
 
 /**
